kobold_discord_bot/requirements.txt
discord.py>=2.3.0
requests>=2.31.0
Flask>=3.0.0
aiohttp>=3.9.0
langdetect>=1.0.9
deep-translator>=1.11.4
